services/api_service/requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Database
sqlalchemy[asyncio]
psycopg[binary,pool]
alembic

# Configuration
pydantic-settings
email-validator

# Authentication
python-jose[cryptography]
bcrypt
python-multipart

# Testing
pytest
pytest-asyncio
httpx

